Feat: Add logic to categorize characters
- Move CharacterCreationForm under form folder
