+++ src/modules/Onboarding/Form/index.ts
@@ -1,0 +1,1 @@
+export * from './CharacterCreationForm.client';
