+++ src/modules/Onboarding/index.ts
@@ -1,3 +1,3 @@
 export * from './CharacterCreation.type';
-export * from './CharacterCreationForm.client';
 export * from './CharacterCreationSteps.client';
+export * from './Form/CharacterCreationForm.client';
